Rebase Brave's changes on top of c3e428cc
Tested with:

 * Visual Studio 2022
 * Windows SDK 10.0.22621.0
 * WTL 10.0.10320
 * Python 2.7
 * pywin32 224
 * SCons 1.3.1
 * Google Software Construction Toolkit 0.9.1
 * Go 1.14.6 (64 bit)
 * Google Protocol Buffers 3.17.3

The following tests fail on Windows 11. They either also fail upstream, or
already failed in our fork before this rebase.

 * TimeTest.RFC822TimeParsing
 * UserRightsTest.UserIsLoggedOnInteractively
 * UtilsTest.IsUserLoggedOn
 * OmahaCustomizationTest.IsInternalUser
 * PingTest.BuildOmahaPing
 * PingTest.SendInProcess
 * PingTest.PersistAndSendPersistedPings
 * WebServicesClientTest.Send
 * WebServicesClientTest.SendForcingHttps
 * WebServicesClientTest.SendWithCustomHeader
 * WebServicesClientTest.SendStringWithCustomHeader
 * DownloadManagerUserTest.DownloadApp_MultiplePackagesInOneApp
 * InstallManagerInstallAppMachineTest.InstallApp_MsiInstallerSucceeds
 * InstallManagerInstallAppMachineTest.InstallApp_MsiInstallerWithArgumentSucceeds
 * InstallerWrapperMachineTest.InstallApp_MsiInstallerSucceeds
 * InstallerWrapperMachineTest.InstallApp_MsiInstallerWithArgumentSucceeds
 * WorkerWithTwoAppsTest.CheckForUpdateAsync_Large
 * WorkerWithTwoAppsTest.DownloadAsyncThenDownloadAndInstallAsync_Large
 * CupEcdsaRequestTest.PostSimpleRequest
 * CupEcdsaRequestTest.PostSimpleRequestHttps
 * NetworkRequestTest.MultipleRequests
 * GoogleUpdateRecoveryTest.FixGoogleUpdate_FileReturned_Machine
 * GoogleUpdateRecoveryTest.FixGoogleUpdate_FileReturned_User
 * GoogleUpdateRecoveryTest.FixGoogleUpdate_FileCollision
 * GoogleUpdateRecoveryTest.VerifyFileSignature_SignedValid
 * GoogleUpdateRecoveryTest.ProductionServerResponseTest
 * IsForeground/WebServicesClientTest.SendUsingCup/0, where GetParam() = false
 * IsForeground/WebServicesClientTest.SendUsingCup/1, where GetParam() = true
 * IsForeground/WebServicesClientTest.SendString/0, where GetParam() = false
 * IsForeground/WebServicesClientTest.SendString/1, where GetParam() = true

(Remember to set OMAHA_PSEXEC_DIR when running tests.)

+++ omaha/mi_exe_stub/x86_encoder/bcj2.cc
@@ -114,7 +114,9 @@ int wmain(int argc, WCHAR* argv[], WCHAR* env[]) {
     DWORD bytes_written = 0;
     if (!::WriteFile(get(file),
                      out.data(),
-                     out.size(),
+                     // This cast is safe because we checked
+                     // out0.size() + ... + out4.size() <= DWORD_MAX above:
+                     static_cast<DWORD>(out.size()),
                      &bytes_written,
                      NULL) ||
         bytes_written != out.size()) {
